move side effects to index.ts
