+++ src/index.ts
@@ -1,6 +1,9 @@
 // tslint:disable-next-line:no-var-requires
 const commander = require("commander");
-import { CommandSections } from "./classes/commandSections";
+import * as child_process from "child_process";
+import * as fs from "fs";
+import * as inquirer from "inquirer";
+import { format, parse } from "path";
 import { StringCompiledHTML } from "./classes/stringCompiledHTML";
 
 // interface of command-line arguments
@@ -11,9 +14,17 @@ interface IArgv {
 function main(): void {
   const args: IArgv = configureCommander();
 
-  const html = StringCompiledHTML.generateFromMarkdownFile(args.file);
-  const commands = CommandSections.generateFromHTML(html);
-  commands.executeCommands();
+  const content = fs.readFileSync(format(parse(args.file)), "utf8");
+  const html = StringCompiledHTML.generateFromMarkdownContent(content);
+  const commands = html.toCommandSections();
+
+  commands.choiceOne((questionCommand, questionName) => {
+    inquirer.prompt([questionCommand]).then(answerCommands => {
+      const cmd = child_process.exec(answerCommands[questionName]);
+      cmd.stdout.pipe(process.stdout);
+      cmd.stderr.pipe(process.stderr);
+    });
+  });
 }
 
 function configureCommander(): IArgv {
